+++ CHANGELOG.md
@@ -10,12 +10,15 @@ Fix:
 
 Features:
   - Rails 5.1.0.rc1 (`potassium` version will follow `Rails` version from now on)
-  - Adds front_end option to include Vue.js and Angular 2 (using `webpacker`)
+  - Adds `front_end` option to include Vue.js and Angular 2 (using `webpacker`)
   - Replaces `bower` by `yarn`
+  - Adds `commoner` option to enable support for ES6 (using `sprockets-commoner`)
 
 Bugfixes:
   - Force `DelayedJobs` migration to use 4.2 `ActiveRecord` class
   - Fix custom `ActiveAdmin` footer
+
+[sprockets-commoner]: https://github.com/Shopify/sprockets-commoner
 
 ## 3.0.0